fixed combine between books-routes and main
