+++ .env
@@ -1,3 +1,6 @@
+# App
+APP_PORT=3000
+
 # OAuth Tokens
 ACCESS_TOKEN_EXP=900
 REFRESH_TOKEN_EXP=21600
