+++ controllers/reportControllers.ts
@@ -1,8 +1,8 @@
+import { Request, Response } from 'express';
 import { BookDTO } from '../dtos/bookDTO';
 import ReportBookDTO from '../dtos/reportBookDTO';
 import ReportBorrowersDTO from '../dtos/reportBorrowersDTO';
 import { ReportRepository } from '../repositories/reportRepository';
-import { Request, Response } from 'express';
 
 export class ReportController {
     private reportRepo = new ReportRepository();
