+++ index.ts
@@ -1,16 +1,16 @@
 import express from "express";
+import dotenv from "dotenv";
 import bookRouter from "./routes/bookRouter";
 import { pool } from "./db/dbConnection";
 import { createTables } from "./db/migrations/createTables";
-import dotenv from "dotenv";
 
 dotenv.config();
 const app = express();
+const port = process.env.APP_PORT!;
 
 app.use(express.json());
 app.use("/", bookRouter);
 
-const port = process.env.PORT || 3000;
 app.listen(port, async () => {
   console.log(`Server started on port ${port}`);
   await pool
