+++ repositories/borrowerRepository.ts
@@ -1,4 +1,4 @@
-import  query  from '../db/db'
+import { pool } from "../db/dbConnection";
 import BorrowerDTO from '../dtos/borrowerDTO';
 
 export class BorrowerRepository{
@@ -6,7 +6,7 @@ export class BorrowerRepository{
         const queryText = 'SELECT borrower_id, first_name, last_name, email FROM borrowers';
 
         try{
-            const result = await query(queryText);
+            const result = await pool.query(queryText);
             return result.rows.map((row)=> new BorrowerDTO(row.first_name, row.last_name, row.email, row.borrower_id));
         }
         catch(err){
@@ -19,7 +19,7 @@ export class BorrowerRepository{
         const queryText = 'SELECT first_name, last_name, email FROM borrowers WHERE borrower_id = $1';
         const values = [id];
         try{
-            const result = await query(queryText, values);
+            const result = await pool.query(queryText, values);
             if(result.rows.length > 0){
                 const {first_name, last_name, email} = result.rows[0];
                 return new BorrowerDTO(first_name, last_name, email, id);
@@ -34,7 +34,7 @@ export class BorrowerRepository{
         const queryText = 'SELECT first_name, last_name, email FROM borrowers WHERE email = $1';
         const values = [email];
         try{
-            const result = await query(queryText, values);
+            const result = await pool.query(queryText, values);
             if(result.rows.length > 0){
                 const {first_name, last_name, email} = result.rows[0];
                 return new BorrowerDTO(first_name, last_name, email);
@@ -49,7 +49,7 @@ export class BorrowerRepository{
         const queryText = 'INSERT INTO borrower (first_name, last_name, email) VALUES ($1, $2, $3) RETURNING id_borrower';
         const values = [newBorrower.firstName, newBorrower.lastName, newBorrower.email];
         try{
-            const result = await query(queryText, values);
+            const result = await pool.query(queryText, values);
             newBorrower.id = result.rows[0].borrower_id;
             return newBorrower;
         }
@@ -61,7 +61,7 @@ export class BorrowerRepository{
         const queryText = 'UPDATE borrower SET first_name = $1, last_name = $2, email = $3 WHERE id_borrower = $4';
         const values = [newBorrower.firstName, newBorrower.lastName, newBorrower.email, newBorrower.id];
         try{
-            await query(queryText, values);
+            await pool.query(queryText, values);
             console.log("Borrower updated");
         }catch(err){
             throw new Error(`Error while updating borrower`);
@@ -72,7 +72,7 @@ export class BorrowerRepository{
         const queryText = 'DELETE FROM borrowers WHERE borrower_id = $1';
         const values = [id];
         try{
-            await query(queryText, values);
+            await pool.query(queryText, values);
             console.log("Borrower deleted");
         }catch(err){
             throw new Error(`Error while deleting borrower`);
