+++ repositories/reportRepository.ts
@@ -1,4 +1,4 @@
-import query from "../db/db";
+import { pool } from "../db/dbConnection";
 import {BookDTO} from "../dtos/bookDTO";
 import {BorrowerDTO} from "../dtos/borrowerDTO";
 import ReportBookDTO from "../dtos/reportBookDTO";
@@ -11,7 +11,7 @@ export class ReportRepository{
     getAllBooks = async () => {
         const queryText = 'SELECT * FROM books';
         try {
-            const result = await query(queryText);
+            const result = await pool.query(queryText);
            const reportAllBooks = new ReportBookDTO();
            result.rows.forEach((row) => {
                 const book = new BookDTO(row.isbn, row.title, row.author, row.year, row.isavailable);
@@ -27,7 +27,7 @@ export class ReportRepository{
     getAllBorrowed = async () => {
         const queryText = 'SELECT * FROM books WHERE isavailable = false';
         try {
-            const result = await query(queryText);
+            const result = await pool.query(queryText);
            const reportBorrowedBooks = new ReportBookDTO();
            result.rows.forEach((row) => {
                 const book = new BookDTO(row.isbn, row.title, row.author, row.year, row.isavailable);
@@ -43,7 +43,7 @@ export class ReportRepository{
     getAllBorrowers = async () => {
         const queryText = 'SELECT * FROM borrower';
         try {
-            const result = await query(queryText);
+            const result = await pool.query(queryText);
             const reportBorrower = new ReportBorrowersDTO();
             result.rows.forEach((row) => {
                 const borrower = new BorrowerDTO(row.id_borrower, row.first_name, row.last_name, row.email);
@@ -59,7 +59,7 @@ export class ReportRepository{
     getAllOverdue = async () => {
         const queryText = 'SELECT bb.ISBN, b.title, b.author_id, b.publication_year, b.is_available FROM books_borrowers bb INNER JOIN books b ON bb.ISBN = b.ISBN WHERE bb.borrowing_date < bb.due_date';
         try {
-            const result = await query(queryText);
+            const result = await pool.query(queryText);
             const reportOverdueBooksDTO = new ReportBookDTO();
             result.rows.forEach((row) => {
                 const book = new BookDTO(row.isbn, row.title, row.author_id, row.publication_year, row.is_available);
