+++ repositories/userRepository.ts
@@ -1,11 +1,11 @@
-import  query  from '../db/db';
+import { pool } from "../db/dbConnection";
 import UserDTO from '../dtos/userDTO';
 
 export class UserRepository {
     async getAll() {
         const queryText = 'SELECT id_user, first_name, last_name, email FROM user';
         try {
-            const result = await query(queryText);
+            const result = await pool.query(queryText);
             return result.rows.map((row) => new UserDTO(row.id_user, row.first_name, row.last_name, row.email, row.hashed_pass));
         }
         catch (err) {
@@ -16,7 +16,7 @@ export class UserRepository {
         const queryText = 'SELECT first_name, last_name, email FROM user WHERE id_user = $1';
         const values = [id];
         try {
-            const result = await query(queryText, values);
+            const result = await pool.query(queryText, values);
             if (result.rows.length > 0) {
                 const { first_name, last_name, email, hashed_pass } = result.rows[0];
                 return new UserDTO(id, first_name, last_name, email, hashed_pass);
@@ -31,7 +31,7 @@ export class UserRepository {
         const queryText = 'INSERT INTO user (first_name, last_name, email) VALUES ($1, $2, $3) RETURNING id_user';
         const values = [user.firstName, user.lastName, user.email];
         try {
-            const result = await query(queryText, values);
+            const result = await pool.query(queryText, values);
             user.id = result.rows[0].id_user;
             return user;
         }
@@ -43,7 +43,7 @@ export class UserRepository {
         const queryText = 'UPDATE user SET first_name = $1, last_name = $2, email = $3 WHERE id_user = $4';
         const values = [user.firstName, user.lastName, user.email, user.id];
         try {
-            await query(queryText, values);
+            await pool.query(queryText, values);
             console.log("User updated");
         }
         catch (err) {
@@ -54,7 +54,7 @@ export class UserRepository {
         const queryText = 'DELETE FROM user WHERE id_user = $1';
         const values = [id];
         try {
-            await query(queryText, values);
+            await pool.query(queryText, values);
             console.log("User deleted");
         }
         catch (err) {
