+++ routers/userRouter.ts
@@ -1,6 +1,5 @@
-import { UserController } from "../controllers/userController";
 import { Router } from "express";
-
+import { UserController } from "../controllers/userController";
 
 const userController = new UserController();
