--- src/server.ts
@@ -1,11 +1,0 @@
-require('dotenv').config();
-import express from 'express';
-
-const app = express();
-const port = 3000;
-
-app.use(express.json());
-
-app.listen(port, () => {
-    console.log(`Server running on port ${port}`)
-});
